Create own go project for filterScript
Rename filterScript to multichain-feed
Structure go code to be able to easily exand the multichain-feed with other hooks/sideEffects
