+++ blockchain/src/multichain-feed/notificationWatcher.js
